Add day 3
Slightly more interesting. Haven't used regex in Uiua before, so good
learning experience there. The do/don't stuff really stumped me for
a while until I was trying to find a way to use scan for it and saw
it could be used with group/box to create groups of do/don't arrays,
which immediately made it easy to combine with rows and fold.