Add day 4
I think part one was harder than part two here, just because there are
more variants to check (i.e., both cardinal and diagonal) as opposed
to one shape that's easy to verify. So, checking that both diagonals
in part two are "SAM" in either direction is straight-forward, checking
for both cardinals and diagonals is quite a bit more tedious.