Add day 17 part 1
Day 17 part 1 is pretty straight-forward as described and fun. Writing
little VMs is always a relaxing time, so was nice to have one to
write. Part 2 I haven't worked out a pattern for yet, so it's up in
the air. Since the operands are all fixed, I have a feeling working
backward from the output instruction since there's no branching in the
code except at the end and it always jumps to zero) is the way to go,
but it's late and I'm tired.

Also update Uiua to a nearby 0.14.0 RC (mainly for subscripts,
which have made certain things easier to write without excess
parentheses). Existing code has been reformatted for it.