Day 17: add part 2 solution in Uiua
More (probably misuse) of path to do a search for all numbers that
would reproduce the VM's code. This could probably be made faster by
doing a custom search but that's a bit more tedious. The amount of
boxing in this gives me a headache too, mostly to keep `path` happy
since it'll otherwise complain about shapes due to each node being
an array of octets (could operate off the or'd number but this was
handy for debugging). So, not fast or optimized, and using the VM steps
instead of the inlined code is slower, but does entire the test-2 input
works, so at lesat it works sort of generally.