Add day 25 part 1
For some reason had a lot of trouble with part 1, but not sure why
yet. Think it might've been a side effect of the size of records
(treating them as 6 with all sums -1 instead of 7) but not sure
yet. Should poke around the input file some more later since the sizes
in the description seem misleading.