Add day 14
Not the fastest, but I'm not as adept at this as others might be. My
part 2 solution was just to log results and look for runs of bots that
looked like they'd made shapes. After that just adjusted the check
for whether the tree was found based on the likelihood of multiple
runs of bots in order, which happens to be correct for this case. I'm
sure there's a maths approach to this that requires zero time, but
that'd probably be difficult for me to work out without knowing where
to start.