Add day 17 part 2 spreadsheet
I haven't gotten around to porting this to Uiua but my day 17 part 2
solution was just a Numbers spreadsheet. I want to eventually translate
it to a breadth-first search looking for 3-bit sequences that build
up the output (3-bit because modulo 8) since that was how it ended
up working in practice. (I don't have the piece of paper I used for
tracking matches, though some remnants are in the spreadsheet I think
from before I gave up and used paper for organization).