Add day 12 part 1
I originally did this a while ago but with `group` instead of
`partition` and assumed it wouldn't be feasible without writing a
flood fill of some kind in Uiua, which I didn't want to do. Turns out
`partition` does exactly what I need, so I'm a doofus. Still kind of
annoyed with myself over that.